requirements.txt
fastapi[all]
uvicorn
pymongo
requests
python-dotenv
fuzzywuzzy[speedup]
pytest
